+++ devicemanagement/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Updating the test plan for resetSOC tests.
 - Added NOPARALLEL label for tests which should not be run in parallel regression.
+- [SW-15160] Added resetSOC stress tests.
 ### Changed
 - [SW-14288] Enabled back setAndGetModuleVoltage test.
 - Updated deviceApi version in Conan requirements.